Closes https://github.com/tamasgal/UnROOT.jl/issues/136 (?)

Use LibDeflate to get a ~30% speedup in zlib decompression. The trick is
```julia
                # Need to ignore the 0x78 0xXX zlib header
                deleteat!(rawbytes, 1:2)
                decompress!(Decompressor(), uncomp_data, rawbytes, uncompbytes)
```

On a 2016 MacBook Pro:
```julia
julia> using UnROOT; const tf = LazyTree(ROOTFile("Run2012BC_DoubleMuParked_Muons.root"), "Events");

# original (zlib.h)
julia> @btime sum(tf.nMuon) # 1.087 s (12162 allocations: 544.53 MiB)
julia> @btime sum(length,tf.Muon_pt) # 5.523 s (18914 allocations: 2.99 GiB)

# this PR (LibDeflate)
julia> @btime sum(tf.nMuon) # 719.714 ms (12162 allocations: 544.50 MiB)
julia> @btime sum(length,tf.Muon_pt) # 4.195 s (18914 allocations: 2.99 GiB)
```

And I get a similar improvement factor on a `x86_64 AMD EPYC 7702P` machine.