@Moelf 2021-11-25
Can we remove CodecZlib then? And also another place wher this is used