@aminnj 2021-11-25
> Can we remove CodecZlib then? And also sync the other place wher this is used

Based on coverage/unit test files not hitting the else statement, I would say 'yes', but I'm not sure about other ROOT files in the wild that might need the CodecZlib stuff