@Moelf 2021-11-26
>If you can hang into the Compressor object instead of creating and destroying it at every compression, it should be even faster.

maybe we should do:
```julia
const _DefPool = [Decompressor() for _ = 1:Threads.nthreads()]

decompress!(_DefPool[Threads.threadid()], uncomp_data, rawbytes, uncompbytes)
```