@jakobnissen 2021-11-26
I just released LibDeflate 0.4.0 (will be in the General Regisry in a few minutes) with support for zlib files.