@aminnj 2021-11-26
> maybe we should do:
> 
> ```julia
> const _DefPool = [Decompressor() for _ = 1:Threads.nthreads()]
> 
> decompress!(_DefPool[Threads.threadid()], uncomp_data, rawbytes, uncompbytes)
> ```

Profiling tells me that `Decompressor()` initialization is <0.2% of the actual basket decompression time, so it's probably not worth it. Out of curiosity, I tried using those two lines and got a segfault:
```julia
julia> using UnROOT ; const tf = LazyTree(ROOTFile("Run2012BC_DoubleMuParked_Muons.root"),"Events")
signal (11): Segmentation fault: 11
in expression starting at REPL[1]:1
deflate_decompress_bmi2 at /Users/namin/.julia/artifacts/7cec4195c0422c4f94c45d6319a2c452d6848da7/lib/libdeflate.0.dylib (unknown line)
libdeflate_deflate_decompress at /Users/namin/.julia/artifacts/7cec4195c0422c4f94c45d6319a2c452d6848da7/lib/libdeflate.0.dylib (unknown line)
_unsafe_decompress! at /Users/namin/.julia/packages/LibDeflate/JYL8P/src/LibDeflate.jl:99 [inlined]
unsafe_decompress! at /Users/namin/.julia/packages/LibDeflate/JYL8P/src/LibDeflate.jl:126 [inlined]
decompress! at /Users/namin/.julia/packages/LibDeflate/JYL8P/src/LibDeflate.jl:158
decompress_datastreambytes at /Users/namin/.julia/dev/UnROOT/src/types.jl:194
```