@Moelf 2021-11-26
We can probably switch to new LibDeflate api and confirm performance gain retains and merge this?