@tamasgal 2021-11-27
This is awesome! Anything else in your mind regarding this PR or can we merge? I'll try it this week with my analysis. Currently in heavy processing mode 🙈