@aminnj 2021-11-27
We can always refine the decompression api further in another pr, so feel free to merge! 😃